[Sonic Origins] Fix crash with S3K Peel Out code
Fixes the crash that would occur when encountering the Marble Garden Zone Act 2 boss as Amy alone, whether in the main game or in Boss Rush, with the aforementioned code active.

The crash is caused by the Tails's `Player::Create` function being called before its character ID is set, leading the function to think the entity is Sonic, or the default character ID. From there, it then attempts to access Sonic's assets/data, which are not loaded in when playing as Amy, leading to the game crashing by attempting to access a null pointer. To fix this, the patch makes it so that Tails's `Create` function isn't called until only after his character ID is set.